requirements.txt
# Core dependencies for Student Bank Account Chatbot
python-dotenv>=0.19.0
google-generativeai>=0.5.0

# Optional: For enhanced terminal experience
colorama>=0.4.4